SpSe-categories-ui: move SettingsManager and CategoriesRepository to data package
