+++ shared/src/commonMain/kotlin/ru/sbx/spend_sense/di/Modules.kt
@@ -2,13 +2,13 @@ package ru.sbx.spend_sense.di
 
 import org.koin.core.module.Module
 import org.koin.dsl.module
+import ru.sbx.spend_sense.data.CategoriesRepository
+import ru.sbx.spend_sense.data.storage.SettingsManager
 import ru.sbx.spend_sense.platform.DeviceInfo
-import ru.sbx.spend_sense.presentation.categories.data.CategoriesRepository
 import ru.sbx.spend_sense.presentation.categories.list.compose.CategoriesViewModel
 import ru.sbx.spend_sense.presentation.common.ui.calendar.DatePickerViewModel
 import ru.sbx.spend_sense.presentation.root.RootViewModel
 import ru.sbx.spend_sense.presentation.settings.SettingsViewModel
-import ru.sbx.spend_sense.storage.SettingsManager
 
 expect val platformModule: Module
