+++ shared/src/commonMain/kotlin/ru/sbx/spend_sense/presentation/categories/list/compose/CategoriesViewModel.kt
@@ -4,11 +4,11 @@ import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.launch
 import kotlinx.datetime.LocalDateTime
+import ru.sbx.spend_sense.data.CategoriesRepository
 import ru.sbx.spend_sense.extentions.now
 import ru.sbx.spend_sense.presentation.base.BaseViewModel
 import ru.sbx.spend_sense.presentation.categories.creation.CreateCategoryData
 import ru.sbx.spend_sense.presentation.categories.creation.toCategory
-import ru.sbx.spend_sense.presentation.categories.data.CategoriesRepository
 import ru.sbx.spend_sense.presentation.categories.model.CategoryContract
 
 class CategoriesViewModel(
