+++ shared/src/commonMain/kotlin/ru/sbx/spend_sense/presentation/root/RootViewModel.kt
@@ -2,10 +2,10 @@ package ru.sbx.spend_sense.presentation.root
 
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
+import ru.sbx.spend_sense.data.storage.SettingsManager
 import ru.sbx.spend_sense.presentation.base.BaseViewModel
 import ru.sbx.spend_sense.presentation.root.model.AppTab
 import ru.sbx.spend_sense.presentation.root.model.RootContract.State
-import ru.sbx.spend_sense.storage.SettingsManager
 
 class RootViewModel(
     settingsManager: SettingsManager
