+++ shared/src/commonMain/kotlin/ru/sbx/spend_sense/presentation/settings/SettingsViewModel.kt
@@ -2,10 +2,10 @@ package ru.sbx.spend_sense.presentation.settings
 
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
+import ru.sbx.spend_sense.data.storage.SettingsManager
 import ru.sbx.spend_sense.platform.DeviceInfo
 import ru.sbx.spend_sense.presentation.base.BaseViewModel
 import ru.sbx.spend_sense.presentation.settings.SettingsContract.State
-import ru.sbx.spend_sense.storage.SettingsManager
 
 class SettingsViewModel(
     deviceInfo: DeviceInfo,
